moved is_text_including_words from intents to entities
